Remove jquery dep. Remove GroupedActionsTable and consolidate tables into single component
