+++ package.json
@@ -2,7 +2,7 @@
   "author": "Dataminr",
   "name": "dataminr-react-components",
   "description": "Collection of reusable React Components and utility functions",
-  "version": "0.8.5",
+  "version": "0.9.0",
   "keywords": [
     "react-component"
   ],
@@ -19,7 +19,6 @@
   "dependencies": {
     "d3": "4.2.6",
     "flux": "2.1.1",
-    "jquery": "2.1.4",
     "lodash": "3.10.1"
   },
   "peerDependencies": {
