--- src/js/table/BasicTable.js
@@ -1,4 +1,0 @@
-var React = require('react');
-var BaseTable = require('./BaseTable');
-
-module.exports = React.createClass(BaseTable);
